requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
colorama==0.4.3
discord.py==1.5.0
idna==2.10
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pylint==2.6.0
python-dotenv==0.14.0
six==1.15.0
toml==0.10.1
wrapt==1.12.1
yarl==1.6.0